+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-10-26 03:02:37
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: ae49760260d07b2c70b7bdc4ee45712f50f1d894
+
+## [ForwardWidgets-2kuai] - 2025-10-26 03:02:37
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 28bc6a32846633ac75b7ffa7d48e2851191073d3
+
+## [ForwardWidgets-pack1r] - 2025-10-26 03:02:37
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 528e140ba675a201660c848391eb27e95f7747c1
+
+
 ## [ForwardWidgets-huangxd] - 2025-10-25 02:55:24
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main